Enhance container create re: devices
Allow API clients to provide a formatted string in the Devices.Path field
which the handler will then parse and stat() the device file information
from the filesystem.

The Path is formatted as "source:destination:permissions", and may be:
 - "/dev/xyz:/dev/custom:rwxrwxrwx"
 - "/dev/xyz:/dev/custom" permissions copied from host
 - "/dev/xyz" all fields copied from host and destination is /dev/xyz

Devices.Major and Devices.Minor must both be zero to signal the handler
to do setup the device. Devices may be a mix of pre-populated
LinuxDevice structs and Devices to be populated by the handler.

Signed-off-by: Jhon Honce <jhonce@redhat.com>